requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
flask-paginate==2022.1.8
fonttools==4.40.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.1
multitasking==0.0.11
numpy==1.25.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tornado==6.3.2
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.3.6
xyzservices==2023.5.0
yfinance==0.2.22